% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_participation_network.R
\name{average_participation_networks}
\alias{average_participation_networks}
\title{Average Participation Network Graphs}
\usage{
average_participation_networks(multiyr_graph, annual_graphs, period, years)
}
\arguments{
\item{multiyr_graph}{igraph object for multi-year participation network}

\item{annual_graphs}{a list of igraph objects; annual participation networks for each year in the multi-year network}

\item{period}{early or late season?}

\item{years}{years included in the multi-year network, and covered by the annual_graphs list}
}
\value{
averaged multi-year igraph object, and the averaged vertex data frame used to build it
}
\description{
Extract vertices attributes from the listed graph objects, and
average numeric attributes together to be incorporated into
a new participation network.
}
\examples{
avg_g <- average_participation_networks(multiyr_graph=multiyr_g,annual_graphs=graphs_list,period=c,years=years)[[1]]
}
