% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multiyear_participation_network.R
\name{participation_network_multiyr}
\alias{participation_network_multiyr}
\title{Generate Multi-year Participation Network}
\usage{
participation_network_multiyr(
  tickets,
  pcid_choose = NA,
  years_choose = c(2013, 2014, 2015),
  filter,
  filter_subgraph = FALSE,
  min_vessels = 3,
  min_contribution = 0.1
)
}
\arguments{
\item{tickets}{fish tickets data frame}

\item{pcid_choose}{specify a port group}

\item{years_choose}{a vector of the crab years to build a network for}

\item{filter}{use the \code{min_vessels} and \code{min_contribution} objects to filter the data}

\item{filter_subgraph}{a filtering option from the original function that was turned off for Fisher et al.}

\item{min_vessels}{the minimum number of vessels participating in a fishery for that fishery to be retained in the network}

\item{min_contribution}{the minimum contribution (as a proportion) to total exvessel revenue for a fishery to be retained for a given vessel}
}
\value{
non-confidential fisheries partition network as an igraph object
}
\description{
Create a fisheries participation network based on an
adjacency matrix that contains multiple crab years of
data. Based on original function by Emma Fuller;
edits specified in comments.
}
\examples{
pre_shock <- participation_network_multiyr(close_dat, filter = TRUE, filter_subgraph = FALSE)
}
