% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3yr_networks.R
\name{plot_3yr}
\alias{plot_3yr}
\title{Simple Participation Network Plotting for 3-year Avg Networks}
\usage{
plot_3yr(g, outdir)
}
\arguments{
\item{g}{network as an igraph object}

\item{outdir}{directory to save the png file (relative)}
}
\description{
Create a simple plot of a participation network and write it out
to a file. The plot will have unlabeled, grey nodes and a
circular layout. The Dungeness crab node will be highlighted
in orange. Note that the default file name is based on
function usage to plot the 3-year average participation
networks used in Figure 1 in Fisher et al.
For script Fig1_3yrNetworks.Rmd
}
\examples{
plot_3yr(tmpgraph, outdir=pngdir)
}
